+++ lib/settings.dart
@@ -6,6 +6,8 @@ const settingsThemeUseDark = 'theme_use_dark';
 
 const settingsUpcomingDays = 'filter_next_up_days';
 const settingsTodayInUpcoming = 'filter_today_in_upcoming';
+const settingsDefaultFilter = 'filter_default_selection';
+
 const settingsAutoAddFilter = 'item_auto_add_filter';
 
 
@@ -18,7 +20,16 @@ const defaultSettings = {
   settingsUpcomingDays: 7,
   settingsAutoAddFilter: true,
   settingsTodayInUpcoming: true,
+  settingsDefaultFilter: 'all',
 };
 
 const nextUpDaysMin = 0;
 const nextUpDaysMax = 90;
+
+const settingsDefaultFilterItems = {
+  'all': 'Everything',
+  'due:today': 'Today',
+  'due:upcoming': 'Upcoming',
+  'due:someday': 'Someday',
+  'due:overdue': 'Overdue',
+};