+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [unreleases]
 ### Added
 - Item context menu (right click) for quick item tasks [#62]
+- Add setting for the default item filter on start-up [#56]
 
 ## [1.0.6] - 2022-03-29
 ### Added